+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Semgrep analyzer changelog
 
+## v4.2.2
+- TEMP: Use patched version of `sast-rules` to remove detect-object-injection (!266)
+- Upgrade [sast-rules](https://gitlab.com/gitlab-org/security-products/sast-rules/-/tags/v1.3.12) version `1.3.12` (!266)
+- chore: Drop high-FP rules removal behavior (!266)
+
 ## v4.2.1
 - Upgrade [sast-rules](https://gitlab.com/gitlab-org/security-products/sast-rules/-/tags/v1.3.11) version `1.3.11` (!249)
   - Flawfinder singular rule IDs should include `-1` [sast-rules!147](https://gitlab.com/gitlab-org/security-products/sast-rules/-/merge_requests/147)